app/backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=24.2
psycopg[binary]>=3.2
drf-spectacular>=0.27
python-dotenv>=1.0

pytest>=8.3
pytest-django>=4.8
django-cors-headers
django-filter